+++ hazelcastmq-examples/src/main/java/org/mpilone/hazelcastmq/example/core/TransactionalSend.java
@@ -58,38 +58,42 @@ public class TransactionalSend extends ExampleApp {
     ClusterNode node2 = new ClusterNode(config);
 
     try {
-      HazelcastMQContext session1 = node1.createContext(true);
+      HazelcastMQContext context1 = node1.createContext(true);
 
-      HazelcastMQProducer producer = session1.createProducer();
+      HazelcastMQProducer producer = context1.createProducer();
+      
+      HazelcastMQContext context2 = node2.createContext(false);
+      HazelcastMQConsumer consumer = context2.createConsumer(queueName);
+
+      // Send the first message while in a transaction.
       producer.send(queueName, "Hello World!".getBytes());
-
-      HazelcastMQContext session2 = node2.createContext(false);
-
-      HazelcastMQConsumer consumer = session2.createConsumer(queueName);
 
       // Try to consume the message that was sent in another node in a
       // transaction. This node shouldn't be able to see the message.
       HazelcastMQMessage msg = consumer.receive(1, TimeUnit.SECONDS);
-      if (msg == null) {
-        log.info("Did not get message (expected behavior).");
-      }
-      else {
-        Assert.fail("Something went wrong. A transactional "
-            + "message shouldn't be visible.");
-      }
+      Assert.isNull(msg, "Something went wrong. A transactional message "
+          + "shouldn't be visible.");
 
       // Commit the transaction so all nodes can see the transactional messages.
-      log.info("Committing producer transaction.");
-      session1.commit();
-      session1.close();
+      log.info("Committing producer transaction 1.");
+      context1.commit();
 
       // Try to consume the message again. This time is should be visible.
       msg = consumer.receive(1, TimeUnit.SECONDS);
       Assert.notNull(msg, "Something went wrong. A message should be visible.");
-
       log.info("Got message body: " + new String(msg.getBody()));
 
-      session2.close();
+      // Send another message in another transaction.
+      producer.send(queueName, "Goodbye World!".getBytes());
+      log.info("Committing producer transaction 2.");
+      context1.commit();
+
+      msg = consumer.receive(1, TimeUnit.SECONDS);
+      Assert.notNull(msg, "Something went wrong. A message should be visible.");
+      log.info("Got message body: " + new String(msg.getBody()));
+
+      context1.close();
+      context2.close();
     }
     finally {
       node1.shutdown();